+++ domain/constants.go
@@ -1,0 +1,9 @@
+package domain
+
+var GameWidth float64 = 750
+var GameHeight float64 = 1000
+var LineWidth float64 = 10
+var LineHeight float64 = 200
+var ColorGreen string = "#07B51E"
+var ColorYellow string = "B57E07"
+var ColorRed string = "#B50707"